feat: split settings into categories
