+++ lib/entries_database.dart
@@ -929,4 +929,35 @@ DROP TABLE old_entries;
       }
     }
   }
+
+  void showLoadingStatus(
+      BuildContext context, ValueNotifier<String> statusNotifier) {
+    showDialog(
+      context: context,
+      barrierDismissible: false,
+      builder: (context) {
+        return PopScope(
+          canPop: false,
+          child: Dialog(
+            child: Container(
+              padding: const EdgeInsets.all(20),
+              child: Column(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  CircularProgressIndicator(),
+                  const SizedBox(height: 20),
+                  ValueListenableBuilder<String>(
+                    valueListenable: statusNotifier,
+                    builder: (context, message, child) {
+                      return Text(message, textAlign: TextAlign.center);
+                    },
+                  ),
+                ],
+              ),
+            ),
+          ),
+        );
+      },
+    );
+  }
 }
