+++ lib/l10n/intl_en.arb
@@ -165,6 +165,8 @@
   "@backupErrorDescription": {},
   "restoreErrorDescription": "Failed to restore backup!",
   "@restoreErrorDescription": {},
+  "settingsBackupRestoreTitle": "Backup & Restore",
+  "@settingsBackupRestoreTitle": {},
   "settingsBackup": "Backup",
   "@settingsBackup": {},
   "settingsRestore": "Restore",
@@ -189,6 +191,10 @@
   "@settingsImport": {},
   "settingsImportFromAnotherApp": "Import From Another App",
   "@settingsImportFromAnotherApp": {},
+  "settingsTranslateCallToAction": "Everyone should have access to a journal!",
+  "@settingsTranslateCallToAction": {},
+  "settingsHelpTranslate": "Help Translate",
+  "@settingsHelpTranslate": {},
   "importLogs": "Import Logs",
   "@importLogs": {},
   "importImages": "Import Images",
@@ -221,6 +227,10 @@
   "@licenseGPLv3": {},
   "settingsSourceCode": "Source Code",
   "@settingsSourceCode": {},
+  "settingsMadeWithLove": "Made with ❤️",
+  "@settingsMadeWithLove": {},
+  "settingsConsiderSupporting": "consider supporting",
+  "@settingsConsiderSupporting": {},
   "tagMoodTitle": "Mood",
   "@tagMoodTitle": {}
 }
